pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ulurn-av <ulurn@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
flake8 = "^7.1.0"
mypy = "^1.10.0"
black = "^24.4.2"
isort = "^5.13.2"
setuptools = "^70.1.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.3"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.4.0"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = 'settings.py|tests'

[tool.isort]
profile = "black"
line_length = 120
skip = ["settings.py", "migrations", "tests"]